+++ app/src/main/java/com/fibelatti/photowidget/configure/PhotoWidgetConfigureState.kt
@@ -11,7 +11,6 @@ data class PhotoWidgetConfigureState(
     val isProcessing: Boolean = true,
     val cropQueue: List<LocalPhoto> = emptyList(),
     val messages: List<Message> = emptyList(),
-    val markedForDeletion: Set<String> = emptySet(),
     val hasEdits: Boolean = false,
 ) {