+++ app/src/main/java/com/fibelatti/photowidget/di/PhotoWidgetModule.kt
@@ -6,6 +6,7 @@ import androidx.room.Room
 import coil.ImageLoader
 import coil.disk.DiskCache
 import coil.memory.MemoryCache
+import com.fibelatti.photowidget.widget.data.ExcludedWidgetPhotoDao
 import com.fibelatti.photowidget.widget.data.PendingDeletionWidgetPhotoDao
 import com.fibelatti.photowidget.widget.data.PhotoWidgetDatabase
 import com.fibelatti.photowidget.widget.data.PhotoWidgetOrderDao
@@ -45,6 +46,11 @@ object PhotoWidgetModule {
     fun photoPendingDeletionPhotoDao(
         photoWidgetDatabase: PhotoWidgetDatabase,
     ): PendingDeletionWidgetPhotoDao = photoWidgetDatabase.pendingDeletionWidgetPhotoDao()
+
+    @Provides
+    fun excludedPhotoDao(
+        photoWidgetDatabase: PhotoWidgetDatabase,
+    ): ExcludedWidgetPhotoDao = photoWidgetDatabase.excludedWidgetPhotoDao()
 
     @Provides
     @Singleton