+++ app/src/main/java/com/fibelatti/photowidget/model/WidgetPhotos.kt
@@ -1,0 +1,6 @@
+package com.fibelatti.photowidget.model
+
+data class WidgetPhotos(
+    val current: List<LocalPhoto>,
+    val excluded: List<LocalPhoto>,
+)